+++ tvb/interfaces/web/templates/genshi/visualizers/pse_isocline/view.html
